Added testing dependency, added HomesteadConfig class test placeholder, and renamed the other tests
