+++ composer.json
@@ -8,6 +8,10 @@
 		"symfony/console": "~2.0",
 		"symfony/yaml": "3.0.*@dev"
 	},
+    "require-dev": {
+        "phpunit/phpunit": "4.7.*@dev",
+        "mikey179/vfsStream": "~1"
+    },
 	"license": "MIT",
 	"authors": [
 		{
